fly - migrate observations to postgres
